requirements.txt
python-environ==0.4.54
selenium==4.18.1